Many changes and bug fixes
Tries to bring the type names more in line with ghcjs-base (so JSVal
instead of JSValueRef).  In the native version JSVal is a ForeignPtr
(where as JSValueRef is a Ptr).

Among the bug fixes are some important reference counting fixes for the
native version.  Referenced values are now protected from collection by
JavaScriptCore (the protection ends when a Haskell ForeignPtr finalizer
runs).