Bug fixes and doctest fixes
The tests have been broken for a while and this should fix them and
a couple of bugs to do with the handling of null strings.  Not super
keen on allowing null strings, but if you make one in JS land
it we should not crash.  Perhaps we can change it in a future version
to throw an exception or make all string functions to return Maybe.