Include ghcjs-base compatible modules
Adds GHC implementations of GHCJS.Marshal, JavaScript.Array.Internal
and JavaScript.Object.Internal.

Most of the default PFromJSVal and PToJSVal instances cannot be
implemented (because we need to know the JS context), but there are
many instances in ghcjs-dom that are.

Also adds modules like Language.Javascript.JSaddle.Array and .Foreign.
These are not quite the same as the ghcjs-base any ghcjs-base modules
because they have some IO versions instead of pure ones.