Fix Chrome issue and async forked threads
Chrome does not allow ws messages to arrive during xhr.  This
means that if there is a sync pending when an async callback happens
the next sync in the callback itself hangs (as it waits for the
first).  This is reasolved by giving sync callbacks their own
syncState MVar.

This change also adds code that should make threads forked from
a sync callback behave correctly (using the async send function
to send to JavaScript).