Fix Firefox compatibility
It turns out that the original code was relying on V8 specific scoping
behavior. The value of the argument `initialSyncDepth` was being set
to `1` even when the other arguments inside `runBatch` were being
evaluated and eventually calling `runBatch` again. So if I were do
make a rough trace:

1. <some function>
2. call to `runBatch` with `initialSyncDepth` set to `1`
3. evaluation of `firstBath` argument of `runBatch`
4. <some code>
5. another call to `runBatch` with no `initialSyncDepth` set
   ^ here is where the difference manifests, in Firefox the value of
   `initialSyncDepth` as seen by the nested call to `runBatch` will be
   `undefined`. However in V8 it will be `1` as the call to `runBatch`
   in 2 is still in progress technically.

This difference would lead to desync between the server side and the
frontend JS code, which manifests as "spinning" in a tight loop as the
frontend is trying to fetch an event that doesn't exist.

Signed-off-by: magic_rb <magic_rb@redalder.org>